Start to build out ECIES
This patch starts to build the functions necessary to run ECIES, as defined by Ethereum (or more realistically: as defined by the other implementations of it, because it's underspecified in all of the documentation). There are two goals for this repo:

1. To make a clear and clear implementation of ECIES
2. To make a set of standard modules in Extereum so that cryptography can performed in one place (and backends can be swapped out universally).

I believe (so far) that if we get this to work, this will be the cleanest implementation of ECIES (versus Go-Ethereum and Pyethereum, at least). This module has a variety of tests that have no yet been made to pass, and then they'll be a second stage trying to get the results to work with DevP2P.